Moved organizationDto to it's dto folder And renamed createOrganization to dto type naming
