+++ backend/Tasque.BLL/MappingProfiles/OrganizationProfile.cs
@@ -1,12 +1,6 @@
 ﻿using AutoMapper;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Tasque.Core.Common.DTO;
+using Tasque.Core.Common.DTO.Organization;
 using Tasque.Core.Common.Entities;
-using Tasque.Core.Common.PartialModels;
 
 namespace Tasque.Core.BLL.MappingProfiles
 {
@@ -14,7 +8,7 @@ namespace Tasque.Core.BLL.MappingProfiles
     {
         public OrganizationProfile()
         {
-            CreateMap<CreateOrganization, Organization>();
+            CreateMap<CreateOrganizationDto, Organization>();
             CreateMap<Organization, OrganizationDto>().ReverseMap();
         }
     }
