+++ backend/Tasque.Core.Common/DTO/Organization/CreateOrganizationDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Tasque.Core.Common.DTO.Organization;
+
+public class CreateOrganizationDto
+{
+    public string? Name { get; set; }
+
+    public int AuthorId { get; set; }
+}
