+++ backend/Tasque.Core.Common/DTO/Organization/OrganizationDto.cs
@@ -1,0 +1,10 @@
+﻿namespace Tasque.Core.Common.DTO.Organization;
+
+public class OrganizationDto
+{
+    public int Id { get; set; }
+
+    public string Name { get; set; } = null!;
+
+    public int AuthorId { get; set; }
+}
