--- backend/Tasque.Core.Common/DTO/PartialModels/CreateOrganization.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Tasque.Core.Common.PartialModels
-{
-    public class CreateOrganization
-    {
-        public string? Name { get; set; }
-
-        public int AuthorId { get; set; }
-    }
-}
